@@ -2,7 +2,7 @@ import { ListItemText, MenuItem } from '@material-ui/core'
 import Select, { SelectProps } from '@material-ui/core/Select'
 import { FieldSetting, GenericContent } from '@sensenet/default-content-types'
 import { Query } from '@sensenet/query'
-import React = require('react')
+import React from 'react'
 
 /**
  * Props object for the SelectFieldP Component
